duplicate KeyModifier for Scala 2 and 3 to pass mima
